backend/requirements.txt
# Core
Django==4.2.10
wagtail==5.2.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2

# Wagtail Extensions
wagtail-localize==1.5.1
wagtailmedia==0.14.4
wagtail-storages==1.0.0

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1
djoser==2.2.2

# API Documentation
drf-yasg==1.21.7

# File Upload & Processing
Pillow==10.2.0
django-storages==1.14.2

# Server
gunicorn==21.2.0
uvicorn==0.24.0
whitenoise==6.6.0

# Development
django-debug-toolbar==4.2.0
python-dotenv==1.0.0

# Export/Import
django-import-export==3.3.1

# Rich Text Editor
django-ckeditor==6.7.0

# Validation
django-phonenumber-field[phonenumbers]==7.3.0

# Date/Time handling
python-dateutil==2.8.2

# Required by Wagtail
django-modelcluster>=6.0
django-taggit>=4.0
draftjs-exporter>=2.1.7
html5lib>=1.1
beautifulsoup4>=4.9.3
anyascii>=0.3.1

# Celery and Redis
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
django-celery-beat==2.5.0

# Additional packages
django-guardian==2.4.0
django-health-check==3.17.0
django-ratelimit==4.1.0
argon2-cffi==23.1.0